requirements.txt
django >6.0,<6.1
python-decouple
pillow
djangorestframework
markdown
django-filter